UMA Fixes
Fixed UCE Character Creation UMA prefab to now have the correct panels and display UMA information when UCE UMA Integration is installed.

Fixed UCE UMA Integration with the archer having the bow in the wrong hand and when changing it was causing a issue with the shield mount, reworked the mounts to fix it. Also made an adjustment to the UMA.NetworkManagerMMO to load uma properly on character preview.

Fixed a minor issue in UCE Extra Skillbar that causes it to disappear when UCE UMA Integration was installed.